+++ src/plugins/intel_cpu/src/nodes/scaled_attn.cpp
@@ -169,7 +169,7 @@ struct MHAKernel {
         auto Hk = present_key.size(1);
         size_t h_each_group_len = H / Hk;
         if (d_scale == 0.0F) {
-            d_scale = 1.0F / sqrt(head_size);
+            d_scale = 1.0F / sqrtf(static_cast<float>(head_size));
         }
 
         auto k_stride_s = present_key.stride(3);
@@ -532,7 +532,7 @@ struct MHAKernel<ScaledDotProductAttention::KT_ONEDNN, T> {
                     float d_scale = 0.0F) {
         auto head_size = query.size(3);
         if (d_scale == 0.0F) {
-            d_scale = 1.0F / sqrt(head_size);
+            d_scale = 1.0F / sqrtf(static_cast<float>(head_size));
         }
 
         prepare_brgemm_prim(strm, query, present_key, present_value, has_out_transpose);
@@ -748,7 +748,7 @@ struct MHAKernel<ScaledDotProductAttention::KT_MLAS, float> {
         size_t h_each_group_len = H / h_group_num;
 
         if (d_scale == 0.0F) {
-            d_scale = 1.0F / sqrt(head_size);
+            d_scale = 1.0F / sqrtf(static_cast<float>(head_size));
         }
         auto k_stride_s = present_key.stride(3);
 
@@ -800,7 +800,7 @@ struct MHAKernel<ScaledDotProductAttention::KT_MLAS, float> {
             if (k_stride_s == 1) {
                 mlas_sgemm("N",
                            "T",
-                           m_cnt,
+                           static_cast<int64_t>(m_cnt),
                            kv_len,
                            head_size,
                            1.0F,
@@ -815,7 +815,7 @@ struct MHAKernel<ScaledDotProductAttention::KT_MLAS, float> {
             } else {
                 mlas_sgemm("N",
                            "N",
-                           m_cnt,
+                           static_cast<int64_t>(m_cnt),
                            kv_len,
                            head_size,
                            1.0F,
@@ -847,7 +847,7 @@ struct MHAKernel<ScaledDotProductAttention::KT_MLAS, float> {
             }
             mlas_sgemm("N",
                        "N",
-                       m_cnt,
+                       static_cast<int64_t>(m_cnt),
                        head_size_v,
                        kv_len,
                        1.0F,
